Explicitly define :sync_name default (issue #695)
Despite :sync_name being a string type argument, in Mac OS X / a few versions
of Ruby reported, the argument's value gets interpreted as nil when not given
on the command line.

This sets the default for the argument to an empty string so that when .empty?
is called on it, it won't produce NoMethodError the way it would for nil