Rename default-sync to shortexample-sync
While reading the documentation I was under the impression that `default-sync` followed some type of hierarchy, for example if I defined `foo-sync` I would expect it to use all settings defined in `foo-sync` plus `default-sync`, like a object merge between the two, that's not the case, it's just the name of the volume. To avoid this potential confusion I am proposing renaming it to `shortexample-sync`.