Remove risky dependency availability memoizations
I've removed memoization of availability check of dependencies that docker-sync likes to install. This is because it's very easy to get in a loop of: Check? No? Install. Check? Still no? Install. This happens for example with unison, as this check is called twice, even though docker-sync installed it in between.

I would prefer to remove this memoization from every dependency class, even the ones where we don't install anything, but I didn't want to do more invasive changes as I'm not familiar with project enough.