Update link in message regarding `dest` removal
When a value for `dest` is set, a message is displayed informing that it
has been removed. This message contains a link pointing to wiki as a
reference. This link is outdated though.

This commit replaces the outdated link with a link to the same section
in the docs that are currently being used (readthedocs).